:art: format: query, command 분리를 위한 패키지 및 코드 구조 변경
