+++ backend/application/common/build.gradle.kts
@@ -1,6 +1,4 @@
 dependencies {
-    implementation(project(":core"))
-
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
