+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/SignUpCommand.kt
@@ -1,0 +1,9 @@
+package com.flavorite.application.common.command
+
+data class SignUpCommand(
+    val email: String,
+    val username: String,
+    val address: String,
+    val password: String,
+    val roles: List<String>
+)
