+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/domain/User.kt
@@ -1,0 +1,16 @@
+package com.flavorite.application.common.command.domain
+
+import com.flavorite.application.common.command.domain.enums.UserRole
+import com.flavorite.application.common.command.domain.vo.Address
+import com.flavorite.application.common.command.domain.vo.Email
+import com.flavorite.application.common.command.domain.vo.Password
+import com.flavorite.application.common.command.domain.vo.Username
+
+class User(
+    val userId: Long,
+    val username: Username,
+    val email: Email,
+    val password: Password,
+    val address: Address,
+    val roles: List<UserRole>
+)
