+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/domain/vo/Password.kt
@@ -1,0 +1,6 @@
+package com.flavorite.application.common.command.domain.vo
+
+data class Password(
+    private val value: String
+) {
+}
