+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/exception/AlreadyUserException.kt
@@ -1,0 +1,6 @@
+package com.flavorite.application.common.command.exception
+
+class AlreadyUserException(
+    val errorCode: ErrorCode
+) : RuntimeException(errorCode.message) {
+}
