+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/exception/NotFoundUserException.kt
@@ -1,0 +1,5 @@
+package com.flavorite.application.common.command.exception
+
+class NotFoundUserException(
+    val errorCode: ErrorCode
+) : RuntimeException(errorCode.message)
