+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/exception/NotMatchedUserRoleException.kt
@@ -1,0 +1,3 @@
+package com.flavorite.application.common.command.exception
+
+class NotMatchedUserRoleException(errorCode: ErrorCode) : RuntimeException()
