+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/port/FindUserPort.kt
@@ -1,0 +1,9 @@
+package com.flavorite.application.common.command.port
+
+import com.flavorite.application.common.command.domain.User
+
+interface FindUserPort {
+
+    fun selectMemberBy(email: String): User?
+
+}
