+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/port/SignUpMemberPort.kt
@@ -1,0 +1,8 @@
+package com.flavorite.application.common.command.port
+
+import com.flavorite.application.common.command.SignUpCommand
+
+interface SignUpMemberPort {
+
+    fun execute(command: SignUpCommand)
+}
