+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/usecases/SignInUseCase.kt
@@ -1,0 +1,6 @@
+package com.flavorite.application.common.command.usecases
+
+interface SignInUseCase {
+
+    fun execute(command: SignInCommand)
+}
