+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/command/usecases/SignUpUseCase.kt
@@ -1,0 +1,8 @@
+package com.flavorite.application.common.command.usecases
+
+import com.flavorite.application.common.command.SignUpCommand
+
+interface SignUpUseCase {
+
+    fun execute(command: SignUpCommand)
+}
