--- backend/application/common/src/main/kotlin/com/flavorite/application/common/dto/command/SignInCommand.kt
@@ -1,8 +1,0 @@
-package com.flavorite.application.common.dto.command
-
-data class SignInCommand(
-    val email: String,
-    val password: String
-) {
-
-}
