--- backend/application/common/src/main/kotlin/com/flavorite/application/common/dto/command/SignUpCommand.kt
@@ -1,9 +1,0 @@
-package com.flavorite.application.common.dto.command
-
-data class SignUpCommand(
-    val email: String,
-    val username: String,
-    val address: String,
-    val password: String,
-    val roles: List<String>
-)
