--- backend/application/common/src/main/kotlin/com/flavorite/application/common/dto/value/MemberDto.kt
@@ -1,10 +1,0 @@
-package com.flavorite.applicatoin.dto.value
-
-data class MemberDto(
-    val userId: Long,
-    val email: String,
-    val username: String,
-    val address: String,
-    val password: String,
-) {
-}
