--- backend/application/common/src/main/kotlin/com/flavorite/application/common/dto/value/SelectUsersQuery.kt
@@ -1,7 +1,0 @@
-package com.flavorite.applicatoin.dto.value
-
-data class SelectUsersQuery(
-    val userIds: List<Long>
-) {
-
-}
