--- backend/application/common/src/main/kotlin/com/flavorite/application/common/exception/AlreadyUserException.kt
@@ -1,6 +1,0 @@
-package com.flavorite.application.common.exception
-
-class AlreadyUserException(
-    val errorCode: ErrorCode
-) : RuntimeException(errorCode.message) {
-}
