--- backend/application/common/src/main/kotlin/com/flavorite/application/common/exception/NotFoundUserException.kt
@@ -1,5 +1,0 @@
-package com.flavorite.application.common.exception
-
-class NotFoundUserException(
-    val errorCode: ErrorCode
-) : RuntimeException(errorCode.message)
