--- backend/application/common/src/main/kotlin/com/flavorite/application/common/port/SignUpMemberPort.kt
@@ -1,8 +1,0 @@
-package com.flavorite.application.common.port
-
-import com.flavorite.application.common.dto.command.SignUpCommand
-
-interface SignUpMemberPort {
-
-    fun execute(command: SignUpCommand)
-}
