--- backend/application/common/src/main/kotlin/com/flavorite/application/common/port/SimpleMemberPort.kt
@@ -1,12 +1,0 @@
-package com.flavorite.application.common.port
-
-import com.flavorite.application.common.dto.value.SecurityMemberDto
-import com.flavorite.applicatoin.dto.value.MemberDto
-
-interface SimpleMemberPort {
-
-    fun selectMemberBy(email: String): MemberDto?
-
-    fun selectSecurityMemberBy(email: String): SecurityMemberDto?
-
-}
