+++ backend/application/common/src/main/kotlin/com/flavorite/application/common/query/port/UserQueryPort.kt
@@ -1,0 +1,9 @@
+package com.flavorite.application.common.query.port
+
+import com.flavorite.application.common.query.dto.SecurityMemberDto
+
+interface UserQueryPort {
+
+    fun selectSecurityMemberBy(email: String): SecurityMemberDto?
+
+}
