--- backend/application/common/src/main/kotlin/com/flavorite/application/common/usecases/SignUpUseCase.kt
@@ -1,8 +1,0 @@
-package com.flavorite.application.common.usecases
-
-import com.flavorite.application.common.dto.command.SignUpCommand
-
-interface SignUpUseCase {
-
-    fun execute(command: SignUpCommand)
-}
