--- backend/core/build.gradle.kts
@@ -1,9 +1,0 @@
-dependencies {
-    testImplementation(platform("org.junit:junit-bom:5.9.1"))
-    testImplementation("org.junit.jupiter:junit-jupiter")
-
-}
-
-tasks.named("bootJar") {
-    enabled = false
-}
