--- backend/core/src/main/kotlin/com/flavorite/core/enum/UserRole.kt
@@ -1,5 +1,0 @@
-package com.flavorite.core.enum
-
-enum class UserRole {
-    USER, ADMIN
-}
