--- backend/core/src/main/kotlin/com/flavorite/core/vo/Address.kt
@@ -1,6 +1,0 @@
-package com.flavorite.core.vo
-
-data class Address(
-    private val roadName: String
-) {
-}
