--- backend/core/src/main/kotlin/com/flavorite/core/vo/Email.kt
@@ -1,14 +1,0 @@
-package com.flavorite.core.vo
-
-data class Email(
-    private val value: String
-) {
-
-    init {
-        validate(value)
-    }
-
-    private fun validate(value: String) {
-        if(!value.contains("@")) throw RuntimeException("잘못된 형식의 이메일입니다.")
-    }
-}
