--- backend/core/src/main/kotlin/com/flavorite/core/vo/Password.kt
@@ -1,6 +1,0 @@
-package com.flavorite.core.vo
-
-data class Password(
-    private val value: String
-) {
-}
