--- backend/core/src/main/kotlin/com/flavorite/core/vo/Username.kt
@@ -1,6 +1,0 @@
-package com.flavorite.core.vo
-
-data class Username(
-    private val value: String
-) {
-}
