+++ backend/presentation/common-api/src/main/kotlin/com/flavorite/presentation/common/api/AuthenticateController.kt
@@ -1,11 +1,10 @@
 package com.flavorite.presentation.common.api
 
-import com.flavorite.application.common.dto.command.SignUpCommand
-import com.flavorite.application.common.usecases.SignUpUseCase
+import com.flavorite.application.common.command.SignUpCommand
+import com.flavorite.application.common.command.usecases.SignUpUseCase
 import com.flavorite.presentation.common.request.CommonRequest
 import com.flavorite.security.components.SecurityPasswordEncoder
 import jakarta.validation.Valid
-import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -22,10 +21,6 @@ class AuthenticateController(
         signUpUseCase.execute(request.toDtoWithPasswordEncrypt(encodedPassword))
     }
 
-    @GetMapping("v1/api/sign/in")
-    fun signIn(): String {
-        return "login"
-    }
 }
 
 private fun CommonRequest.SignUp.toDtoWithPasswordEncrypt(encodedPassword: String): SignUpCommand = SignUpCommand(
