+++ backend/settings.gradle.kts
@@ -1,6 +1,5 @@
 rootProject.name = "backend"
 include("bootstrap")
-include("core")
 include("infrastructure:database")
 include("infrastructure:clients")
 include("application:common")
